Simplified separate traits files into single traits.rs
