--- src/traits/mod.rs
@@ -1,23 +1,0 @@
-/*
-Project: thormotion
-GitHub: https://github.com/MillieFD/thormotion
-Author: Amelia Fraser-Dale
-License: BSD 3-Clause "New" or "Revised"
-Filename: traits/mod.rs
-Description: This file defines the traits module, which contains submodules for different traits.
-Each trait contains functions which can be called from Thorlabs devices that implement the trait.
----------------------------------------------------------------------------------------------------
-Notes:
-*/
-
-mod chan_enable_state;
-mod hub;
-mod motor;
-mod thorlabs_device;
-mod unit_conversion;
-
-pub use chan_enable_state::ChanEnableState;
-pub use hub::Hub;
-pub use motor::Motor;
-pub use thorlabs_device::{MsgFormat, ThorlabsDevice};
-pub use unit_conversion::UnitConversion;
